+++ eslintConfig.js
@@ -38,7 +38,7 @@ module.exports = {
     'max-len': [
       'error',
       {
-        code: 100,
+        code: 120,
         ignoreComments: true,
       },
     ],