+++ packages/plugin-image/web/src/image-viewer.jsx
@@ -285,7 +285,7 @@ class ImageViewer extends React.Component {
         onClick={this.handleClick}
         className={itemClassName}
         onKeyDown={this.onKeyDown}
-        ref={e => this.handleRef(e)}
+        ref={this.handleRef}
         onContextMenu={this.handleContextMenu}
         {...accesibilityProps}
       >