+++ CHANGELOG.md
@@ -15,13 +15,18 @@
     Changes that have landed in master but are not yet released.
     Click to see more.
   </summary>
+
+### :bug: Bug Fix
+- `image`
+  - [#1467](https://github.com/wix-incubator/rich-content/pull/1467) image caption aligned correctly in rtl
+
 </details>
 <hr/>
 
 ## 7.16.10 (Aug 19, 2020)
 ### :bug: Bug Fix
 - `editor`
-  - [#1466](https://github.com/wix-incubator/rich-content/pull/1466) default formatting toolbars change - remove first divider 
+  - [#1466](https://github.com/wix-incubator/rich-content/pull/1466) default formatting toolbars change - remove first divider
 
 
 ## 7.16.9 (Aug 18, 2020)